pyproject.toml
[project]
name = "aircraft-detector-classificator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.3",
    "fiftyone>=1.14.0",
    "hydra-core>=1.3.2",
    "mlflow>=3.10.1",
    "onnx>=1.21.0",
    "onnxruntime>=1.24.4",
    "opencv-python>=4.13.0.92",
    "optimum[onnxruntime]>=2.1.0",
    "pycocotools>=2.0.11",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.26",
    "tritonclient[http]>=2.67.0",
    "ultralytics>=8.4.33",
    "uvicorn>=0.44.0",
]

[dependency-groups]
dev = [
    "ruff>=0.15.9",
]

[tool.ruff]
line-length = 100
target-version = "py313"
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "mlruns",
    "notebooks",
]

[tool.ruff.lint]
# E, W: стандарты PEP8 (pycodestyle)
# F:    базовые ошибки (pyflakes)
# I:    сортировка импортов (isort)
select = ["E", "F", "I", "W"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
